requirements.txt
gunicorn
flask
google-generativeai
supabase
dotenv
functools
pygame
time
